Initial grunp engine
Started creating a system that uses a generic entities object to act as
main player and enemies objects.

Might be a little overzealous.

Gotta find an elegant way to handle some simple collision detection and
maybe change the namespace for update() for like a updateMain() or
something so that custom stuff can be coded into update()

After that, should be able to move onto gameplay.

Or I could switch back to Phaser -_-