Version .1 of optiWink implementation
optiWink is the name for the game framework I'm making for grunp so that
I can easily create more html5 shet in the future.

The goal is to make it pretty simple so that I don't need to use Phaser
for everything and so creating these simple things won't require me to
write so much garbage that ends up cluttering the global scope.

Currently new instances of optiWink creates a canvas that tries to take
up the whole page. It needs a way to accept state objects to handle
init() and update() functions per state object.

Or something.

Yeah.